Extract folder browser source management
